requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-ai==0.0.12
python-dotenv==1.0.0
supabase==2.3.4
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
